fix: cannot have multiple TypeScriptCode in the same scope
