+++ src/private/utils.ts
@@ -1,3 +1,4 @@
+import { IConstruct } from 'constructs';
 import { BuildOptions, Platform, TransformOptions } from '../esbuild-types';
 
 export function isEsbuildError(error: unknown): boolean {
@@ -22,3 +23,16 @@ export function defaultPlatformProps(options?: BuildOptions | TransformOptions):
 
   return {};
 }
+
+const assetIds = new WeakMap<IConstruct, number>();
+export const uniqueAssetId = (scope: IConstruct, name: string) => {
+  const nextId = (assetIds.get(scope) ?? 0) + 1;
+  assetIds.set(scope, nextId);
+
+  // Only one asset per scope, skip the id
+  if (nextId === 1) {
+    return name;
+  }
+
+  return `${name}${nextId}`;
+};
